recorder/pyproject.toml
[project]
name = "recorder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.12.2",
]

[tool.ruff]

[tool.ruff.format]
